Error checks code refactors generate docs (#95)
* Regenerate docs

* The first segment to an empty Message should always be MSH

* Add missing files

* Misc code refactor

* Composer update

* Bump PHP to 8.1

* Refactor test

* Run actions on pushing to all branches
