+++ composer.json
@@ -14,7 +14,7 @@
     "minimum-stability": "dev",
     "prefer-stable" : true,
     "require": {
-        "php": "^8.0",
+        "php": ">=8.1",
         "ext-mbstring": "*"
     },
     "require-dev": {
