+++ docs/Connection.md
@@ -4,9 +4,7 @@ Usage:
 ```php
 $connection = new Connection('127.0.0.1', 5002);
 $req = new Message();
-// .
-
-.. set some request attributes
+// ... set some request attributes
 $response = $connection->send($req);
 $response->toString(); // Read ACK message from remote
 ```
@@ -36,6 +34,7 @@ End of message signal for HL7 server. Defaults to \034\015.
 |------|-------------|
 |[__construct](#connection__construct)|Creates a connection to a HL7 server, or throws exception when a connection could not be established.|
 |[__destruct](#connection__destruct)||
+|[getSocket](#connectiongetsocket)||
 |[send](#connectionsend)|Sends a Message object over this connection.|
 
 
@@ -46,7 +45,7 @@ End of message signal for HL7 server. Defaults to \034\015.
 **Description**
 
 ```php
-public __construct (string $host, string $port)
+public __construct (string $host, int $port, int $timeout)
 ```
 
 Creates a connection to a HL7 server, or throws exception when a connection could not be established. 
@@ -55,14 +54,20 @@ Creates a connection to a HL7 server, or throws exception when a connection coul
 
 **Parameters**
 
-* `(string) $host`
-: Host to connect to  
-* `(string) $port`
-: Port to connect to  
+* `(string) $host`: Host to connect to  
+* `(int) $port`: Port to connect to  
+* `(int) $timeout`: Connection timeout  
 
 **Return Values**
 
 `void`
+
+
+**Throws Exceptions**
+
+
+`\HL7ConnectionException`
+
 
 <hr />
 
@@ -72,7 +77,7 @@ Creates a connection to a HL7 server, or throws exception when a connection coul
 **Description**
 
 ```php
-public __destruct (void)
+ __destruct (void)
 ```
 
  
@@ -87,6 +92,31 @@ public __destruct (void)
 
 `void`
 
+
+<hr />
+
+
+### Connection::getSocket  
+
+**Description**
+
+```php
+ getSocket (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
 <hr />
 
 
@@ -95,7 +125,7 @@ public __destruct (void)
 **Description**
 
 ```php
-public send (\Message $req, string $responseCharEncoding)
+public send (string $responseCharEncoding, bool $noWait)
 ```
 
 Sends a Message object over this connection. 
@@ -104,14 +134,23 @@ Sends a Message object over this connection.
 
 **Parameters**
 
-* `(\Message) $req`
 * `(string) $responseCharEncoding`
 : The expected character encoding of the response.  
+* `(bool) $noWait`
+: Do no wait for ACK. Helpful for building load testing tools...  
 
 **Return Values**
 
-`\Message`
+`void`
 
+
+**Throws Exceptions**
+
+
+`\HL7ConnectionException`
+
+
+`\HL7Exception`
 
 
 <hr />
