+++ docs/Messages/ACK.md
@@ -13,7 +13,7 @@ Aranyasen\HL7\Message
 | Name | Description |
 |------|-------------|
 |[setAckCode](#acksetackcode)|Set the acknowledgement code for the acknowledgement.|
-|[setErrorMessage](#ackseterrormessage)|Set the error message for the acknowledgement.|
+|[setErrorMessage](#ackseterrormessage)|Set the error message for acknowledgement.|
 
 ## Inherited methods
 
@@ -21,15 +21,27 @@ Aranyasen\HL7\Message
 |------|-------------|
 |__construct|Constructor for Message. Consider using the HL7 factory to obtain a message instead.|
 |addSegment|Append a segment to the end of the message|
+|getFirstSegmentInstance|Return the first segment with given name in the message|
 |getSegmentAsString|Get the segment identified by index as string, using the messages separators.|
 |getSegmentByIndex|Return the segment specified by $index.|
 |getSegmentFieldAsString|Get the field identified by $fieldIndex from segment $segmentIndex.|
+| [getSegmentIndex](https://secure.php.net/manual/en/aranyasen\hl7\message.getsegmentindex.php) | - |
 |getSegments|Return an array containing all segments in the right order.|
 |getSegmentsByName|Return an array of all segments with the given name|
+|hasSegment|Check if given segment is present in the message object|
 |insertSegment|Insert a segment.|
+| [isAdt](https://secure.php.net/manual/en/aranyasen\hl7\message.isadt.php) | - |
+| [isEmpty](https://secure.php.net/manual/en/aranyasen\hl7\message.isempty.php) | - |
+| [isOrm](https://secure.php.net/manual/en/aranyasen\hl7\message.isorm.php) | - |
+| [isOru](https://secure.php.net/manual/en/aranyasen\hl7\message.isoru.php) | - |
+| [isSiu](https://secure.php.net/manual/en/aranyasen\hl7\message.issiu.php) | - |
+| [removeSegment](https://secure.php.net/manual/en/aranyasen\hl7\message.removesegment.php) | - |
 |removeSegmentByIndex|Remove the segment indexed by $index.|
+|removeSegmentsByName|Remove given segment|
+|resetSegmentIndices|Reset index attribute of each given segment, so when those are added the indices start from 1|
 |segmentToString|Convert Segment object to string|
 |setSegment|Set the segment on index.|
+|toFile|Write HL7 to a file|
 |toString|Return a string representation of this message.|
 
 
@@ -39,7 +51,7 @@ Aranyasen\HL7\Message
 **Description**
 
 ```php
-public setAckCode (string $code, string $msg)
+public setAckCode (string $code, string|null $msg)
 ```
 
 Set the acknowledgement code for the acknowledgement. 
@@ -52,13 +64,12 @@ This denotes: accept, general error and reject respectively. The ACK module will
 
 * `(string) $code`
 : Code to use in acknowledgement  
-* `(string) $msg`
+* `(string|null) $msg`
 : Acknowledgement message  
 
 **Return Values**
 
-`boolean`
-
+`void`
 
 
 <hr />
@@ -72,7 +83,7 @@ This denotes: accept, general error and reject respectively. The ACK module will
 public setErrorMessage (string $msg)
 ```
 
-Set the error message for the acknowledgement. 
+Set the error message for acknowledgement. 
 
 This will also set the error code to either AE or CE, depending on the mode of the incoming message. 
 
@@ -84,6 +95,7 @@ This will also set the error code to either AE or CE, depending on the mode of t
 **Return Values**
 
 `void`
+
 
 <hr />
