+++ docs/README.md
@@ -1,16 +1,32 @@
 # Aranyasen\HL7
 
-* [Connection](Connection.md) 
 * [Message](Message.md) 
-* [Messages\ACK](Messages/ACK.md) 
 * [Segment](Segment.md) 
-* [Segments\MSH](Segments/MSH.md) 
-* [Segments\PID](Segments/PID.md) 
-* [Segments\IN1](Segments/IN1.md) 
-* [Segments\IN3](Segments/IN3.md) 
-* [Segments\NTE](Segments/NTE.md) 
-* [Segments\DG1](Segments/DG1.md) 
-* [Segments\OBR](Segments/OBR.md) 
-* [Segments\OBX](Segments/OBX.md) 
-* [Segments\ORC](Segments/ORC.md) 
-* [Segments\PV1](Segments/PV1.md) 
+* [Messages\ACK](Messages\ACK.md) 
+* [Connection](Connection.md) 
+* [Segments\AIG](Segments\AIG.md) 
+* [Segments\AIL](Segments\AIL.md) 
+* [Segments\AIP](Segments\AIP.md) 
+* [Segments\DG1](Segments\DG1.md) 
+* [Segments\EQU](Segments\EQU.md) 
+* [Segments\EVN](Segments\EVN.md) 
+* [Segments\FHS](Segments\FHS.md) 
+* [Segments\FTS](Segments\FTS.md) 
+* [Segments\GT1](Segments\GT1.md) 
+* [Segments\IN1](Segments\IN1.md) 
+* [Segments\IN3](Segments\IN3.md) 
+* [Segments\MRG](Segments\MRG.md) 
+* [Segments\MSA](Segments\MSA.md) 
+* [Segments\MSH](Segments\MSH.md) 
+* [Segments\NK1](Segments\NK1.md) 
+* [Segments\NTE](Segments\NTE.md) 
+* [Segments\OBR](Segments\OBR.md) 
+* [Segments\OBX](Segments\OBX.md) 
+* [Segments\ORC](Segments\ORC.md) 
+* [Segments\PID](Segments\PID.md) 
+* [Segments\PV1](Segments\PV1.md) 
+* [Segments\PV2](Segments\PV2.md) 
+* [Segments\RGS](Segments\RGS.md) 
+* [Segments\SAC](Segments\SAC.md) 
+* [Segments\SCH](Segments\SCH.md) 
+* [Segments\TQ1](Segments\TQ1.md)
