+++ docs/Segments/FHS.md
@@ -1,0 +1,629 @@
+# Aranyasen\HL7\Segments\FHS  
+
+FHS: File Header Segment
+Ref: https://hl7-definition.caristix.com/v2/HL7v2.3/Segments/FHS
+
+
+
+## Extend:
+
+Aranyasen\HL7\Segment
+
+## Methods
+
+| Name | Description |
+|------|-------------|
+|[getFileControlId](#fhsgetfilecontrolid)||
+|[getFileCreationDateTime](#fhsgetfilecreationdatetime)||
+|[getFileEncodingCharacters](#fhsgetfileencodingcharacters)||
+|[getFileFieldSeparator](#fhsgetfilefieldseparator)||
+|[getFileHeaderComment](#fhsgetfileheadercomment)||
+|[getFileNameId](#fhsgetfilenameid)||
+|[getFileRecievingApplication](#fhsgetfilerecievingapplication)||
+|[getFileRecievingFacility](#fhsgetfilerecievingfacility)||
+|[getFileSecurity](#fhsgetfilesecurity)||
+|[getFileSendingApplication](#fhsgetfilesendingapplication)||
+|[getFileSendingFacility](#fhsgetfilesendingfacility)||
+|[getReferenceFileControlId](#fhsgetreferencefilecontrolid)||
+|[setFileControlId](#fhssetfilecontrolid)||
+|[setFileCreationDateTime](#fhssetfilecreationdatetime)||
+|[setFileEncodingCharacters](#fhssetfileencodingcharacters)||
+|[setFileFieldSeparator](#fhssetfilefieldseparator)||
+|[setFileHeaderComment](#fhssetfileheadercomment)||
+|[setFileNameId](#fhssetfilenameid)||
+|[setFileRecievingApplication](#fhssetfilerecievingapplication)||
+|[setFileRecievingFacility](#fhssetfilerecievingfacility)||
+|[setFileSecurity](#fhssetfilesecurity)||
+|[setFileSendingApplication](#fhssetfilesendingapplication)||
+|[setFileSendingFacility](#fhssetfilesendingfacility)||
+|[setReferenceFileControlId](#fhssetreferencefilecontrolid)||
+
+## Inherited methods
+
+| Name | Description |
+|------|-------------|
+|__construct|Create a segment.|
+|clearField|Remove any existing value from the field|
+|getField|Get the field at index.|
+|getFields|Get fields from a segment|
+|getName|Get the name of the segment. This is basically the value at index 0|
+|setField|Set the field specified by index to value.|
+|size|Get the number of fields for this segment, not including the name|
+
+
+
+### FHS::getFileControlId  
+
+**Description**
+
+```php
+ getFileControlId (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FHS::getFileCreationDateTime  
+
+**Description**
+
+```php
+ getFileCreationDateTime (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FHS::getFileEncodingCharacters  
+
+**Description**
+
+```php
+ getFileEncodingCharacters (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FHS::getFileFieldSeparator  
+
+**Description**
+
+```php
+ getFileFieldSeparator (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FHS::getFileHeaderComment  
+
+**Description**
+
+```php
+ getFileHeaderComment (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FHS::getFileNameId  
+
+**Description**
+
+```php
+ getFileNameId (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FHS::getFileRecievingApplication  
+
+**Description**
+
+```php
+ getFileRecievingApplication (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FHS::getFileRecievingFacility  
+
+**Description**
+
+```php
+ getFileRecievingFacility (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FHS::getFileSecurity  
+
+**Description**
+
+```php
+ getFileSecurity (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FHS::getFileSendingApplication  
+
+**Description**
+
+```php
+ getFileSendingApplication (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FHS::getFileSendingFacility  
+
+**Description**
+
+```php
+ getFileSendingFacility (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FHS::getReferenceFileControlId  
+
+**Description**
+
+```php
+ getReferenceFileControlId (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FHS::setFileControlId  
+
+**Description**
+
+```php
+ setFileControlId (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FHS::setFileCreationDateTime  
+
+**Description**
+
+```php
+ setFileCreationDateTime (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FHS::setFileEncodingCharacters  
+
+**Description**
+
+```php
+ setFileEncodingCharacters (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FHS::setFileFieldSeparator  
+
+**Description**
+
+```php
+ setFileFieldSeparator (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FHS::setFileHeaderComment  
+
+**Description**
+
+```php
+ setFileHeaderComment (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FHS::setFileNameId  
+
+**Description**
+
+```php
+ setFileNameId (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FHS::setFileRecievingApplication  
+
+**Description**
+
+```php
+ setFileRecievingApplication (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FHS::setFileRecievingFacility  
+
+**Description**
+
+```php
+ setFileRecievingFacility (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FHS::setFileSecurity  
+
+**Description**
+
+```php
+ setFileSecurity (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FHS::setFileSendingApplication  
+
+**Description**
+
+```php
+ setFileSendingApplication (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FHS::setFileSendingFacility  
+
+**Description**
+
+```php
+ setFileSendingFacility (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FHS::setReferenceFileControlId  
+
+**Description**
+
+```php
+ setReferenceFileControlId (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
