+++ docs/Segments/FTS.md
@@ -1,0 +1,129 @@
+# Aranyasen\HL7\Segments\FTS  
+
+FTS: File Trailer Segment
+Ref: https://hl7-definition.caristix.com/v2/HL7v2.3/Segments/FTS
+
+
+
+## Extend:
+
+Aranyasen\HL7\Segment
+
+## Methods
+
+| Name | Description |
+|------|-------------|
+|[getFileBatchCount](#ftsgetfilebatchcount)||
+|[getFileTrailerComment](#ftsgetfiletrailercomment)||
+|[setFileBatchCount](#ftssetfilebatchcount)||
+|[setFileTrailerComment](#ftssetfiletrailercomment)||
+
+## Inherited methods
+
+| Name | Description |
+|------|-------------|
+|__construct|Create a segment.|
+|clearField|Remove any existing value from the field|
+|getField|Get the field at index.|
+|getFields|Get fields from a segment|
+|getName|Get the name of the segment. This is basically the value at index 0|
+|setField|Set the field specified by index to value.|
+|size|Get the number of fields for this segment, not including the name|
+
+
+
+### FTS::getFileBatchCount  
+
+**Description**
+
+```php
+ getFileBatchCount (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FTS::getFileTrailerComment  
+
+**Description**
+
+```php
+ getFileTrailerComment (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FTS::setFileBatchCount  
+
+**Description**
+
+```php
+ setFileBatchCount (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### FTS::setFileTrailerComment  
+
+**Description**
+
+```php
+ setFileTrailerComment (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
