+++ docs/Segments/MRG.md
@@ -1,0 +1,379 @@
+# Aranyasen\HL7\Segments\MRG  
+
+MRG segment class
+Ref: http://hl7-definition.caristix.com:9010/HL7%20v2.3.1/segment/MRG
+
+
+
+## Extend:
+
+Aranyasen\HL7\Segment
+
+## Methods
+
+| Name | Description |
+|------|-------------|
+|[getPriorAlternatePatientID](#mrggetprioralternatepatientid)||
+|[getPriorAlternateVisitID](#mrggetprioralternatevisitid)||
+|[getPriorPatientAccountNumber](#mrggetpriorpatientaccountnumber)||
+|[getPriorPatientID](#mrggetpriorpatientid)||
+|[getPriorPatientIdentifierList](#mrggetpriorpatientidentifierlist)||
+|[getPriorPatientName](#mrggetpriorpatientname)||
+|[getPriorVisitNumber](#mrggetpriorvisitnumber)||
+|[setPriorAlternatePatientID](#mrgsetprioralternatepatientid)||
+|[setPriorAlternateVisitID](#mrgsetprioralternatevisitid)||
+|[setPriorPatientAccountNumber](#mrgsetpriorpatientaccountnumber)||
+|[setPriorPatientID](#mrgsetpriorpatientid)||
+|[setPriorPatientIdentifierList](#mrgsetpriorpatientidentifierlist)||
+|[setPriorPatientName](#mrgsetpriorpatientname)||
+|[setPriorVisitNumber](#mrgsetpriorvisitnumber)||
+
+## Inherited methods
+
+| Name | Description |
+|------|-------------|
+|__construct|Create a segment.|
+|clearField|Remove any existing value from the field|
+|getField|Get the field at index.|
+|getFields|Get fields from a segment|
+|getName|Get the name of the segment. This is basically the value at index 0|
+|setField|Set the field specified by index to value.|
+|size|Get the number of fields for this segment, not including the name|
+
+
+
+### MRG::getPriorAlternatePatientID  
+
+**Description**
+
+```php
+ getPriorAlternatePatientID (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### MRG::getPriorAlternateVisitID  
+
+**Description**
+
+```php
+ getPriorAlternateVisitID (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### MRG::getPriorPatientAccountNumber  
+
+**Description**
+
+```php
+ getPriorPatientAccountNumber (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### MRG::getPriorPatientID  
+
+**Description**
+
+```php
+ getPriorPatientID (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### MRG::getPriorPatientIdentifierList  
+
+**Description**
+
+```php
+ getPriorPatientIdentifierList (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### MRG::getPriorPatientName  
+
+**Description**
+
+```php
+ getPriorPatientName (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### MRG::getPriorVisitNumber  
+
+**Description**
+
+```php
+ getPriorVisitNumber (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### MRG::setPriorAlternatePatientID  
+
+**Description**
+
+```php
+ setPriorAlternatePatientID (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### MRG::setPriorAlternateVisitID  
+
+**Description**
+
+```php
+ setPriorAlternateVisitID (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### MRG::setPriorPatientAccountNumber  
+
+**Description**
+
+```php
+ setPriorPatientAccountNumber (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### MRG::setPriorPatientID  
+
+**Description**
+
+```php
+ setPriorPatientID (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### MRG::setPriorPatientIdentifierList  
+
+**Description**
+
+```php
+ setPriorPatientIdentifierList (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### MRG::setPriorPatientName  
+
+**Description**
+
+```php
+ setPriorPatientName (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### MRG::setPriorVisitNumber  
+
+**Description**
+
+```php
+ setPriorVisitNumber (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
