+++ docs/Segments/TQ1.md
@@ -1,0 +1,729 @@
+# Aranyasen\HL7\Segments\TQ1  
+
+TQ1 segment class
+Ref: http://hl7-definition.caristix.com:9010/HL7%20v2.5.1/dataType/Default.aspx?version=HL7+v2.5.1&dataType=TQ1
+
+
+
+## Extend:
+
+Aranyasen\HL7\Segment
+
+## Methods
+
+| Name | Description |
+|------|-------------|
+|[getConditionText](#tq1getconditiontext)||
+|[getConjunction](#tq1getconjunction)||
+|[getEndDateTime](#tq1getenddatetime)||
+|[getExplicitTime](#tq1getexplicittime)||
+|[getOccurrenceDuration](#tq1getoccurrenceduration)||
+|[getPriority](#tq1getpriority)||
+|[getQuantity](#tq1getquantity)||
+|[getRelativeTimeAndUnits](#tq1getrelativetimeandunits)||
+|[getRepeatPattern](#tq1getrepeatpattern)||
+|[getServiceDuration](#tq1getserviceduration)||
+|[getSetIdTQ1](#tq1getsetidtq1)||
+|[getStartDateTime](#tq1getstartdatetime)||
+|[getTextInstruction](#tq1gettextinstruction)||
+|[getTotalOccurrences](#tq1gettotaloccurrences)||
+|[setConditionText](#tq1setconditiontext)||
+|[setConjunction](#tq1setconjunction)||
+|[setEndDateTime](#tq1setenddatetime)||
+|[setExplicitTime](#tq1setexplicittime)||
+|[setOccurrenceDuration](#tq1setoccurrenceduration)||
+|[setPriority](#tq1setpriority)||
+|[setQuantity](#tq1setquantity)||
+|[setRelativeTimeAndUnits](#tq1setrelativetimeandunits)||
+|[setRepeatPattern](#tq1setrepeatpattern)||
+|[setServiceDuration](#tq1setserviceduration)||
+|[setSetIdTQ1](#tq1setsetidtq1)||
+|[setStartDateTime](#tq1setstartdatetime)||
+|[setTextInstruction](#tq1settextinstruction)||
+|[setTotalOccurrences](#tq1settotaloccurrences)||
+
+## Inherited methods
+
+| Name | Description |
+|------|-------------|
+|__construct|Create a segment.|
+|clearField|Remove any existing value from the field|
+|getField|Get the field at index.|
+|getFields|Get fields from a segment|
+|getName|Get the name of the segment. This is basically the value at index 0|
+|setField|Set the field specified by index to value.|
+|size|Get the number of fields for this segment, not including the name|
+
+
+
+### TQ1::getConditionText  
+
+**Description**
+
+```php
+ getConditionText (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::getConjunction  
+
+**Description**
+
+```php
+ getConjunction (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::getEndDateTime  
+
+**Description**
+
+```php
+ getEndDateTime (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::getExplicitTime  
+
+**Description**
+
+```php
+ getExplicitTime (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::getOccurrenceDuration  
+
+**Description**
+
+```php
+ getOccurrenceDuration (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::getPriority  
+
+**Description**
+
+```php
+ getPriority (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::getQuantity  
+
+**Description**
+
+```php
+ getQuantity (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::getRelativeTimeAndUnits  
+
+**Description**
+
+```php
+ getRelativeTimeAndUnits (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::getRepeatPattern  
+
+**Description**
+
+```php
+ getRepeatPattern (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::getServiceDuration  
+
+**Description**
+
+```php
+ getServiceDuration (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::getSetIdTQ1  
+
+**Description**
+
+```php
+ getSetIdTQ1 (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::getStartDateTime  
+
+**Description**
+
+```php
+ getStartDateTime (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::getTextInstruction  
+
+**Description**
+
+```php
+ getTextInstruction (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::getTotalOccurrences  
+
+**Description**
+
+```php
+ getTotalOccurrences (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::setConditionText  
+
+**Description**
+
+```php
+ setConditionText (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::setConjunction  
+
+**Description**
+
+```php
+ setConjunction (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::setEndDateTime  
+
+**Description**
+
+```php
+ setEndDateTime (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::setExplicitTime  
+
+**Description**
+
+```php
+ setExplicitTime (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::setOccurrenceDuration  
+
+**Description**
+
+```php
+ setOccurrenceDuration (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::setPriority  
+
+**Description**
+
+```php
+ setPriority (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::setQuantity  
+
+**Description**
+
+```php
+ setQuantity (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::setRelativeTimeAndUnits  
+
+**Description**
+
+```php
+ setRelativeTimeAndUnits (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::setRepeatPattern  
+
+**Description**
+
+```php
+ setRepeatPattern (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::setServiceDuration  
+
+**Description**
+
+```php
+ setServiceDuration (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::setSetIdTQ1  
+
+**Description**
+
+```php
+ setSetIdTQ1 (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::setStartDateTime  
+
+**Description**
+
+```php
+ setStartDateTime (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::setTextInstruction  
+
+**Description**
+
+```php
+ setTextInstruction (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
+
+### TQ1::setTotalOccurrences  
+
+**Description**
+
+```php
+ setTotalOccurrences (void)
+```
+
+ 
+
+ 
+
+**Parameters**
+
+`This function has no parameters.`
+
+**Return Values**
+
+`void`
+
+
+<hr />
+
