+++ src/HL7/Message.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Aranyasen\HL7;
 
 use Aranyasen\Exceptions\HL7Exception;
-use InvalidArgumentException;
 
 /**
  * Class specifying the HL7 message, both request and response.
@@ -18,18 +17,19 @@ use InvalidArgumentException;
  */
 class Message
 {
+    use SegmentManagerTrait;
     use MessageHelpersTrait;
 
     protected array $segments = [];
 
     protected string $segmentSeparator;
     protected bool $segmentEndingBar; # true, if '|' at end of each segment is needed
-    protected $fieldSeparator;
+    protected string $fieldSeparator;
     protected string $componentSeparator;
     protected string $subcomponentSeparator;
     protected string $repetitionSeparator;
     protected string $escapeChar;
-    protected $hl7Version;
+    protected string $hl7Version;
 
     // Split (or not) repeated subfields joined by ~. E.g. if true, parses 3^0~4^1 to [3, '0~4', 1]
     protected bool $doNotSplitRepetition;
@@ -48,7 +48,7 @@ class Message
      * been added to the message will result in setting these values for the message.
      *
      * If the message couldn't be created, for example due to a erroneous HL7 message string, an error is raised.
-     * @param string|null $msgStr
+     * @param string|null $msgString
      * @param array|null $hl7Globals Set control characters or HL7 properties. e.g., ['HL7_VERSION' => '2.5']
      * @param bool $keepEmptySubFields Set this to true to retain empty sub-fields
      * @param bool $resetIndices Reset Indices of each segment to 1.
@@ -56,9 +56,10 @@ class Message
      * @param bool|null $doNotSplitRepetition If true, repeated segments will be in single array instead of sub-arrays.
      *     Note: Since this is non-standard, it may be removed in future!
      * @throws HL7Exception
+     * @deprecated Use HL7 factory class instead. Check readme
      */
     public function __construct(
-        string $msgStr = null,
+        string $msgString = null,
         array $hl7Globals = null,
         bool $keepEmptySubFields = false,
         bool $resetIndices = false,
@@ -82,16 +83,16 @@ class Message
         }
 
         // If no HL7 string is passed to the constructor, nothing else to do
-        if (!$msgStr) {
+        if (!$msgString) {
             return;
         }
 
-        $segments = preg_split("/[\n\r" . $this->segmentSeparator . ']/', $msgStr, -1, PREG_SPLIT_NO_EMPTY);
+        $segments = preg_split("/[\n\r{$this->segmentSeparator}]/", $msgString, -1, PREG_SPLIT_NO_EMPTY);
         $this->setSeparators($segments[0]); // First segment is MSH, the control segment
 
         // Do all segments
         foreach ($segments as $index => $segmentString) {
-            $fields = preg_split("/\\" . $this->fieldSeparator . '/', $segmentString);
+            $fields = preg_split("/\\{$this->fieldSeparator}/", $segmentString);
             $segmentName = array_shift($fields);
 
             foreach ($fields as $j => $field) {
@@ -107,149 +108,6 @@ class Message
 
             $this->addSegment($segment);
         }
-    }
-
-    /**
-     * Append a segment to the end of the message
-     */
-    public function addSegment(Segment $segment): bool
-    {
-        if (count($this->segments) === 0) {
-            $this->resetCtrl($segment);
-        }
-
-        $this->segments[] = $segment;
-
-        return true;
-    }
-
-    /**
-     * Insert a segment.
-     *
-     * @param null|int $index Index where segment is inserted
-     * @throws InvalidArgumentException
-     */
-    public function insertSegment(Segment $segment, int $index = null): void
-    {
-        if ($index > count($this->segments)) {
-            throw new InvalidArgumentException("Index out of range. Index: $index, Total segments: " .
-                                               count($this->segments));
-        }
-
-        if ($index === 0) {
-            $this->resetCtrl($segment);
-            array_unshift($this->segments, $segment);
-        } elseif ($index === count($this->segments)) {
-            $this->segments[] = $segment;
-        } else {
-            $this->segments =
-                array_merge(
-                    array_slice($this->segments, 0, $index),
-                    [$segment],
-                    array_slice($this->segments, $index)
-                );
-        }
-    }
-
-    /**
-     * Return the segment specified by $index.
-     *
-     * Note: Segment count within the message starts at 0.
-     *
-     * @param int $index Index where segment is inserted
-     */
-    public function getSegmentByIndex(int $index): ?Segment
-    {
-        if ($index >= count($this->segments)) {
-            return null;
-        }
-
-        return $this->segments[$index];
-    }
-
-    public function getSegmentIndex(Segment $segment): ?int
-    {
-        foreach ($this->segments as $ii => $value) {
-            if ($value === $segment) {
-                return $ii;
-            }
-        }
-        return null;
-    }
-
-    /**
-     * Return an array of all segments with the given name
-     *
-     * @param string $name Segment name
-     * @return array List of segments identified by name
-     */
-    public function getSegmentsByName(string $name): array
-    {
-        $segmentsByName = [];
-
-        foreach ($this->segments as $seg) {
-            if ($seg->getName() === $name) {
-                $segmentsByName[] = $seg;
-            }
-        }
-
-        return $segmentsByName;
-    }
-
-    /**
-     * Remove the segment indexed by $index.
-     *
-     * If it doesn't exist, nothing happens, if it does, all segments
-     * after this one will be moved one index up.
-     *
-     * @param int $index Index where segment is removed
-     */
-    public function removeSegmentByIndex(int $index): bool
-    {
-        if ($index < count($this->segments)) {
-            array_splice($this->segments, $index, 1);
-        }
-
-        return true;
-    }
-
-    /**
-     * Remove given segment
-     *
-     * @return int Count of segments removed
-     */
-    public function removeSegmentsByName(string $segmentName): int
-    {
-        $count = 0;
-        foreach ($this->getSegmentsByName($segmentName) as $segment) {
-            $this->removeSegmentByIndex($this->getSegmentIndex($segment));
-            $count++;
-        }
-        return $count;
-    }
-
-    /**
-     * Set the segment on index.
-     *
-     * If index is out of range, or not provided, do nothing. Setting MSH on index 0 will re-validate field separator,
-     * control characters and hl7 version, based on MSH(1), MSH(2) and MSH(12).
-     *
-     * @param int $index Index where segment is set
-     * @throws InvalidArgumentException
-     */
-    public function setSegment(Segment $segment, int $index): bool
-    {
-        if (!isset($index) || $index > count($this->segments)) {
-            throw new InvalidArgumentException('Index out of range');
-        }
-
-        if ($index === 0 && $segment->getName() === 'MSH') {
-            $this->resetCtrl($segment);
-        }
-
-        $this->segments[$index] = $segment;
-
-        return true;
     }
 
     /**
@@ -276,27 +134,15 @@ class Message
     }
 
     /**
-     * Return an array containing all segments in the right order.
-     *
-     * @return array List of all segments
-     */
-    public function getSegments(): array
-    {
-        return $this->segments;
-    }
-
-    /**
      * Return a string representation of this message.
      *
      * This can be used to send the message over a socket to an HL7 server. To print to other output, use the $pretty
      * argument as some true value. This will not use the default segment separator, but '\n' instead.
      *
      * @param boolean $pretty Whether to use \n as separator or default (\r).
-     * @return mixed String representation of HL7 message
-     * @access public
      * @throws HL7Exception
      */
-    public function toString(bool $pretty = false)
+    public function toString(bool $pretty = false): string
     {
         if (empty($this->segments)) {
             throw new HL7Exception('Message contains no data. Can not convert to string');
@@ -320,53 +166,6 @@ class Message
         return $message;
     }
 
-    /**
-     * Convert Segment object to string
-     */
-    public function segmentToString(Segment $seg): string
-    {
-        $segmentName = $seg->getName();
-        $segmentString = $segmentName . $this->fieldSeparator;
-        $fields = $seg->getFields(($segmentName === 'MSH' ? 2 : 1));
-
-        foreach ($fields as $field) {
-            if (is_array($field)) {
-                foreach ($field as $index => $value) {
-                    is_array($value)
-                        ? ($segmentString .= implode($this->subcomponentSeparator, $value))
-                        : ($segmentString .= $value);
-
-                    if ($index < (count($field) - 1)) {
-                        $segmentString .= $this->componentSeparator;
-                    }
-                }
-            } else {
-                $segmentString .= $field;
-            }
-
-            $segmentString .= $this->fieldSeparator;
-        }
-
-        return $segmentString;
-    }
-
-    /**
-     * Reset index attribute of each given segment, so when those are added the indices start from 1
-     */
-    public function resetSegmentIndices(): void
-    {
-        $reflector = new \ReflectionClass($this);
-        $segments = glob(dirname($reflector->getFileName()) . '/Segments/*.php');
-
-        // Go through each available segment class and reset its ID
-        foreach ($segments as $file) { // ['OBR', 'PID', 'OBX', 'IN1'...]
-            $className = "Aranyasen\\HL7\\Segments\\" . pathinfo($file, PATHINFO_FILENAME);
-            if (class_exists($className) && method_exists($className, 'resetIndex')) {
-                $className::resetIndex();
-            }
-        }
-    }
-
     private function extractComponentsFromField(string $field, bool $keepEmptySubFields): array|string
     {
         $pregFlags = $keepEmptySubFields
@@ -374,7 +173,7 @@ class Message
             : PREG_SPLIT_NO_EMPTY;
 
         if ((str_contains($field, $this->repetitionSeparator)) && (! $this->doNotSplitRepetition)) {
-            $components = preg_split("/\\" . $this->repetitionSeparator . '/', $field, -1, $pregFlags);
+            $components = preg_split("/\\{$this->repetitionSeparator}/", $field, -1, $pregFlags);
             $fields = [];
             foreach ($components as $index => $component) {
                 $fields[$index] = $this->extractComponentsFromField($component, $keepEmptySubFields);
@@ -383,13 +182,13 @@ class Message
             return $fields;
         }
 
-        $components = preg_split("/\\" . $this->componentSeparator . '/', $field, -1, $pregFlags);
+        $components = preg_split("/\\{$this->componentSeparator}/", $field, -1, $pregFlags);
         foreach ($components as $index => $component) {
-            $subComps = preg_split("/\\" . $this->subcomponentSeparator . '/', $component);
+            $subComponents = preg_split("/\\{$this->subcomponentSeparator}/", $component);
             // Make it a ref or just the value
-            $components[$index] = count($subComps) === 1
-                ? $subComps[0]
-                : $subComps;
+            $components[$index] = count($subComponents) === 1
+                ? $subComponents[0]
+                : $subComponents;
         }
 
         return count($components) === 1
@@ -410,7 +209,7 @@ class Message
 
         [, , $fieldSep, $compSep, $repSep, $esc, $subCompSep, $fieldSepCtrl] = $matches;
 
-        // Check whether field separator is repeated after 4 control characters
+        // Check whether the field separator is used after 4 control characters
         if ($fieldSep !== $fieldSepCtrl) {
             throw new HL7Exception('Not a valid message: field separator invalid', E_USER_ERROR);
         }
@@ -423,21 +222,5 @@ class Message
         $this->subcomponentSeparator = $subCompSep;
         $this->escapeChar = $esc;
         $this->repetitionSeparator = $repSep;
-    }
-
-    private function getSegmentClass(string $segmentName, array $fields, bool $autoIncrementIndices): Segment
-    {
-        // If a class exists for the segment under segments/, (e.g., MSH)
-        $className = "Aranyasen\\HL7\\Segments\\$segmentName";
-        if (!class_exists($className)) {
-            return new Segment($segmentName, $fields);
-        }
-
-        if ($segmentName === 'MSH') {
-            array_unshift($fields, $this->fieldSeparator); # First field for MSH is '|'
-            return new $className($fields);
-        }
-
-        return new $className($fields, $autoIncrementIndices);
     }
 }
